refactor perceptual losses
