+++ beast/rendering/__init__.py
@@ -1,1 +1,1 @@
-"""Rendering utilities: Gaussian splatting and novel-view synthesis losses."""
+"""Rendering utilities: Gaussian splatting renderer."""
